refactor: move upgrade_lts_contract implementation into uaclient module
