+++ uaclient/tests/test_upgrade_lts_contract.py
@@ -3,7 +3,7 @@ import logging
 import mock
 import pytest
 
-from lib.upgrade_lts_contract import process_contract_delta_after_apt_lock
+from uaclient.upgrade_lts_contract import process_contract_delta_after_apt_lock
 
 
 @pytest.mark.parametrize("caplog_text", [logging.DEBUG], indirect=True)
@@ -13,13 +13,15 @@ class TestUpgradeLTSContract:
         new_callable=mock.PropertyMock,
         return_value=False,
     )
-    def test_unattached_noops(self, m_is_attached, capsys, caplog_text):
+    def test_unattached_noops(
+        self, m_is_attached, capsys, caplog_text, FakeConfig
+    ):
         expected_logs = [
             "Check whether to upgrade-lts-contract",
             "Skipping upgrade-lts-contract. Machine is unattached",
         ]
 
-        process_contract_delta_after_apt_lock()
+        process_contract_delta_after_apt_lock(FakeConfig())
 
         assert 1 == m_is_attached.call_count
         out, _err = capsys.readouterr()
@@ -33,10 +35,16 @@ class TestUpgradeLTSContract:
         new_callable=mock.PropertyMock,
         return_value=True,
     )
-    @mock.patch("lib.upgrade_lts_contract.system.get_release_info")
-    @mock.patch("lib.upgrade_lts_contract.system.subp")
+    @mock.patch("uaclient.upgrade_lts_contract.system.get_release_info")
+    @mock.patch("uaclient.upgrade_lts_contract.system.subp")
     def test_upgrade_cancel_when_past_version_not_supported(
-        self, m_subp, m_get_release_info, m_is_attached, capsys, caplog_text
+        self,
+        m_subp,
+        m_get_release_info,
+        m_is_attached,
+        capsys,
+        caplog_text,
+        FakeConfig,
     ):
         m_get_release_info.return_value = mock.MagicMock(series="groovy")
         m_subp.return_value = ("", "")
@@ -47,7 +55,7 @@ class TestUpgradeLTSContract:
         ]
         expected_logs = ["Check whether to upgrade-lts-contract"]
         with pytest.raises(SystemExit) as execinfo:
-            process_contract_delta_after_apt_lock()
+            process_contract_delta_after_apt_lock(FakeConfig())
 
         assert 1 == execinfo.value.code
         assert 1 == m_is_attached.call_count
@@ -64,10 +72,12 @@ class TestUpgradeLTSContract:
         new_callable=mock.PropertyMock,
         return_value=True,
     )
-    @mock.patch("lib.upgrade_lts_contract.system.get_release_info")
-    @mock.patch("lib.upgrade_lts_contract.system.subp")
-    @mock.patch("lib.upgrade_lts_contract.contract.process_entitlements_delta")
-    @mock.patch("lib.upgrade_lts_contract.time.sleep")
+    @mock.patch("uaclient.upgrade_lts_contract.system.get_release_info")
+    @mock.patch("uaclient.upgrade_lts_contract.system.subp")
+    @mock.patch(
+        "uaclient.upgrade_lts_contract.contract.process_entitlements_delta"
+    )
+    @mock.patch("uaclient.upgrade_lts_contract.time.sleep")
     def test_upgrade_contract_when_apt_lock_is_held(
         self,
         m_sleep,
@@ -107,10 +117,10 @@ class TestUpgradeLTSContract:
         ]
 
         with mock.patch(
-            "lib.upgrade_lts_contract.UAConfig",
+            "uaclient.upgrade_lts_contract.UAConfig",
             return_value=FakeConfig(),
         ):
-            process_contract_delta_after_apt_lock()
+            process_contract_delta_after_apt_lock(FakeConfig())
 
         assert 1 == m_is_attached.call_count
         assert 1 == m_get_release_info.call_count
